wip vps init

+++ vps_initial.sh
@@ -1,23 +1,59 @@
-!#/bin/bash
+#!/bin/bash
+usage() {
+  echo "Usage: $0 [-f] [-s] [-n] [-h]"
+  echo "  -s    Include ssh configuration"
+  echo "  -n    Include neovim install"
+  echo "  -f    Include all"
+  echo "  -h    Print this help message"
+  exit 1
+}
+
+while getopts "fsn" $opt do
+    case $opt in
+        f)  
+         SSH=true
+         NEOVIM=true
+         esac
+         ;;
+        s)
+         SSH=true
+         ;;
+        n)
+         NEOVIM=true
+         ;;
+    esac
+done
 
 # Standart update and upgrade and package install
 sudo apt-get update && apt-get upgrade
 sudo apt-get install zip unzip curl wget nginx postgresql redis-server tmux build-essential 
 
-# Install python security (for work)
-sudo apt install libssl-dev libsqlite3-dev libbz2-dev libgdbm-dev libncurses5-dev libncursesw5-dev libreadline-dev zlib1g-dev libffi-dev
-sudo apt install python3 python3-pip
-echo python3 --version >> ./output.file
-echo python3-pip --version >> ./output.file
-
 # Install neovim
-curl -LO https://github.com/neovim/neovim/releases/latest/download/nvim-linux64.tar.gz
-sudo rm -rf /opt/nvim
-sudo tar -C /opt -xzf nvim-linux64.tar.gz
+if [[ $NEOVIM = true ]]; then
+    echo "installing neovim"
+    curl -LO https://github.com/neovim/neovim/releases/latest/download/nvim-linux64.tar.gz
+    sudo rm -rf /opt/nvim
+    sudo tar -C /opt -xzf nvim-linux64.tar.gz
 
-echo export PATH="$PATH:/opt/nvim-linux64/bin" > ~/.bashrc
-echo export PATH="$PATH:/opt/nvim-linux64/bin" > ~/.zshrc
-. ~/.bashrc
-. ~/.zshrc
+    echo export PATH="$PATH:/opt/nvim-linux64/bin" > ~/.bashrc
+    echo export PATH="$PATH:/opt/nvim-linux64/bin" > ~/.zshrc
+    . ~/.bashrc
+    . ~/.zshrc
+fi
 
 # Configure ssh 
+# MAKE SURE YOU HAVE YOUR id_rsa key root dir
+if [[ $SSH = true ]]; then
+    echo "configuring ssh"
+    sudo cp ~/id_rsa ~/.ssh/authorized_keys_fake
+    ssh-keygen -t rsa -b 2048 -f "~/.ssh/id_rsa_fake" -N "" -q
+    sudo tee /etc/ssh/ssh_config.d/custom.false_conf > /dev/null <<EOL
+PasswordAuthentication no
+Port 56432
+Allowusers $(whoami)
+PubkeyAuthentication yes
+X11Forwarding yes
+PermitRootLogin no
+EOL
+#    sudo service ssh restart
+fi
